requirements.txt
matplotlib==3.9.3
numpy==2.1.3
Pillow==11.0.0
scikit_learn==1.5.2
scikit_image==0.24.0
torch_geometric==2.6.1
torchvision==0.20.1
opencv_python==4.10.0.84